# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API feita para registrar niveis e desenvolvedores",
        "contact": {
            "email": "rafael.g.campi@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/niveis": {
            "get": {
                "tags": [
                    "Niveis"
                ],
                "summary": "Lista todos os níveis",
                "operationId": "f983113f46d3765dfcacc06fe03b483c",
                "parameters": [
                    {
                        "name": "nivel",
                        "in": "query",
                        "description": "Nome do nível",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de níveis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Nivel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Niveis"
                ],
                "summary": "Cria um novo nível",
                "operationId": "372073b181193e713a4c2631caf170f8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Nivel"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nível criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Nivel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
        "/api/niveis/{id}": {
            "get": {
                "tags": [
                    "Niveis"
                ],
                "summary": "Exibe um nível específico",
                "operationId": "40938244cfd89332e1a939aa93c953c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do nível",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nível encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Nivel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nível não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Niveis"
                ],
                "summary": "Atualiza um nível específico",
                "operationId": "360c2793e59e977ef778fed3f550babf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do nível",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Nivel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nível atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Nivel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Nível não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Niveis"
                ],
                "summary": "Remove um nível específico",
                "operationId": "95b63514dd90008c4185702e4c5f9ea9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do nível",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Nível removido com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Nível não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Nivel": {
                "description": "Modelo de Nível",
                "properties": {
                    "id": {
                        "description": "ID do nível",
                        "type": "integer"
                    },
                    "nivel": {
                        "description": "Nome do nível",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
...
    }
}